Track fifo usage & clean up resources when finished
Using shared-memory to keep track of number of users
of the named pipe. Remove resources when last one
is finished.

+++ usr/vtllibrary.c
@@ -1309,6 +1309,7 @@ int main(int argc, char *argv[])
 	int ret;
 	long pollInterval = 0L;
 	uint8_t *buf;
+	int fifo_retval;
 
 	int last_state = MHVTL_STATE_UNKNOWN;
 
@@ -1558,6 +1559,14 @@ int main(int argc, char *argv[])
 	if (lunit.fifoname)
 		open_fifo(&lunit.fifo_fd, lunit.fifoname);
 
+	fifo_retval = inc_fifo_count(lunit.fifoname);
+	if (fifo_retval == -ENOMEM) {
+		MHVTL_LOG("shared memory setup failed - exiting...");
+		goto exit;
+	} else if (fifo_retval < 0) {
+		MHVTL_LOG("Failed to set fifo count()...");
+	}
+
 	for (;;) {
 		/* Check for any messages */
 		mlen = msgrcv(r_qid, &r_entry, MAXOBN, my_id, IPC_NOWAIT);
@@ -1621,10 +1630,10 @@ exit:
 	free(buf);
 	if (lunit.fifo_fd) {
 		fclose(lunit.fifo_fd);
-		unlink(lunit.fifoname);
+		if (!dec_fifo_count(lunit.fifoname))
+			unlink(lunit.fifoname);
 		free(lunit.fifoname);
 	}
 
 	exit(0);
 }
-
